[models] Fixed 500 error on empty category #153
Admin panel was returning 500 Internal server error when a Fimrware build was created with empty category but non-empty OS identifier. This occurred because the "clean" function in "AbstractBuild" model was checking for uniqueness of ("category.organization", "os") before other fields were validated hence generating a RelatedObjectDoesNotExist error on Category.
This change skips the uniqueness check if either the os or category is undefined. The empty category is then caught by other validation checks.