[chores] Improvements round 2: merged remove files logic
We already had a method that removed files, it makes no sense to
have 2 methods which perform very similar actions.
I merged the new logic with the old method, mostly overwriting it.
I updated the tests and simplified the code when possible,
reducing redundant comments and unnecessary lines of code.