[fix] Ensured firmware image files get always deleted from filesystem #301
When deleting a firmware build, a firmware category, or an organization,
the delete operation cascades to related models and firmware images are
removed as well.

However, due to how Django handles cascade deletes, the method
AbstractFirmwareImage.delete() is not invoked, causing image files to
linger on the filesystem.

This patch fixes the issue by collecting file paths to remove and
passing them to a background task which deletes the files and cleans up
empty directories.

Fixes #301